+++ src/backends/base/subopt/subopt_debug.cpp
@@ -22,7 +22,7 @@ namespace mrna::md::base {
 SuboptDebug::SuboptDebug(Primary r, Model::Ptr m, DpState dp, SuboptCfg cfg)
     : r_(std::move(r)), m_(std::move(m)), dp_(std::move(dp)), cfg_(cfg) {}
 
-int SuboptDebug::Run(const SuboptCallback& fn) {
+int64_t SuboptDebug::Run(const SuboptCallback& fn) {
   const int N = static_cast<int>(r_.size());
   verify(N < std::numeric_limits<Index>::max(), "RNA too long for suboptimal folding");
 
@@ -62,7 +62,7 @@ int SuboptDebug::Run(const SuboptCallback& fn) {
 
     // If we found a non-finished node, but `finished` is full, and the worst in `finished` is
     // as good as our current node (which is the best in `q`), then we can exit.
-    if (static_cast<int>(finished_.size()) >= cfg_.strucs &&
+    if (static_cast<int64_t>(finished_.size()) >= cfg_.strucs &&
         (--finished_.end())->res.energy <= node.res.energy)
       break;
 
@@ -446,7 +446,7 @@ int SuboptDebug::Run(const SuboptCallback& fn) {
     assert(struc.not_yet_expanded.empty());
     fn(struc.res);
   }
-  return static_cast<int>(finished_.size());
+  return static_cast<int64_t>(finished_.size());
 }
 
 }  // namespace mrna::md::base
